refactor: add shared folder to reuse code by imports
- add interfaces, constants and helpers folders
- add path aliases to the shared folders
- clean up contentScript file
- update app component import paths
