+++ src/app/shared/helpers/async.helper.ts
@@ -1,0 +1,1 @@
+export const sleep = (ms: number) => new Promise((cb: () => void) => setTimeout(cb, ms));
