+++ src/app/shared/interfaces/environment.interface.ts
@@ -1,9 +1,3 @@
-export interface Settings {
-    selected: number;
-    environments: Environment[];
-    token: string;
-}
-
 export interface Environment {
     id: string;
     name: string;
