+++ src/app/shared/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './environment.interface';
+export * from './settings.interface';
