+++ src/app/shared/interfaces/settings.interface.ts
@@ -1,0 +1,7 @@
+import { Environment } from './environment.interface';
+
+export interface Settings {
+    selected: number;
+    environments: Environment[];
+    token: string;
+}
