+++ tsconfig.json
@@ -14,7 +14,12 @@
     "lib": [
       "es2018",
       "dom"
-    ]
+    ],
+    "paths": {
+      "@constants": ["src/app/shared/constants/index"],
+      "@helpers": ["src/app/shared/helpers/index"],
+      "@interfaces": ["src/app/shared/interfaces/index"],
+    }
   },
   "angularCompilerOptions": {
     "fullTemplateTypeCheck": true,
